renamed categories to category-overview component
